--- instrumentation/rocketmq-client-4.8/library/build.gradle.kts
@@ -1,18 +1,0 @@
-plugins {
-  id("otel.library-instrumentation")
-}
-
-dependencies {
-  library("org.apache.rocketmq:rocketmq-client:4.8.0")
-
-  compileOnly("com.google.auto.value:auto-value-annotations")
-  annotationProcessor("com.google.auto.value:auto-value")
-
-  testLibrary("org.apache.rocketmq:rocketmq-test:4.8.0")
-
-  testImplementation(project(":instrumentation:rocketmq-client-4.8:testing"))
-}
-
-tasks.withType<Test>().configureEach {
-  systemProperty("testLatestDeps", findProperty("testLatestDeps") as Boolean)
-}
