--- instrumentation/rocketmq-client-4.8/library/src/main/java/io/opentelemetry/instrumentation/rocketmq/MapSetter.java
@@ -1,21 +1,0 @@
-/*
- * Copyright The OpenTelemetry Authors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-package io.opentelemetry.instrumentation.rocketmq;
-
-import io.opentelemetry.context.propagation.TextMapSetter;
-import org.apache.rocketmq.client.hook.SendMessageContext;
-
-enum MapSetter implements TextMapSetter<SendMessageContext> {
-  INSTANCE;
-
-  @Override
-  public void set(SendMessageContext carrier, String key, String value) {
-    if (carrier == null) {
-      return;
-    }
-    carrier.getMessage().getProperties().put(key, value);
-  }
-}
