+++ instrumentation/rocketmq/rocketmq-clients/rocketmq-client-4.8/library/src/main/java/io/opentelemetry/instrumentation/rocketmqclient/v4_8/ContextAndScope.java
@@ -1,0 +1,26 @@
+/*
+ * Copyright The OpenTelemetry Authors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+package io.opentelemetry.instrumentation.rocketmqclient.v4_8;
+
+import com.google.auto.value.AutoValue;
+import io.opentelemetry.context.Context;
+import io.opentelemetry.context.Scope;
+
+@AutoValue
+abstract class ContextAndScope {
+
+  static ContextAndScope create(Context context, Scope scope) {
+    return new AutoValue_ContextAndScope(context, scope);
+  }
+
+  abstract Context getContext();
+
+  abstract Scope getScope();
+
+  void close() {
+    getScope().close();
+  }
+}
