+++ instrumentation/rocketmq/rocketmq-clients/rocketmq-client-4.8/library/src/main/java/io/opentelemetry/instrumentation/rocketmqclient/v4_8/TextMapExtractAdapter.java
@@ -1,0 +1,25 @@
+/*
+ * Copyright The OpenTelemetry Authors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+package io.opentelemetry.instrumentation.rocketmqclient.v4_8;
+
+import io.opentelemetry.context.propagation.TextMapGetter;
+import javax.annotation.Nullable;
+import org.apache.rocketmq.common.message.MessageExt;
+
+enum TextMapExtractAdapter implements TextMapGetter<MessageExt> {
+  INSTANCE;
+
+  @Override
+  public Iterable<String> keys(MessageExt carrier) {
+    return carrier.getProperties().keySet();
+  }
+
+  @Nullable
+  @Override
+  public String get(@Nullable MessageExt carrier, String key) {
+    return carrier == null ? null : carrier.getProperties().get(key);
+  }
+}
