+++ settings.gradle.kts
@@ -404,9 +404,9 @@ include(":instrumentation:restlet:restlet-2.0:library")
 include(":instrumentation:restlet:restlet-2.0:testing")
 include(":instrumentation:rmi:bootstrap")
 include(":instrumentation:rmi:javaagent")
-include(":instrumentation:rocketmq-client-4.8:javaagent")
-include(":instrumentation:rocketmq-client-4.8:library")
-include(":instrumentation:rocketmq-client-4.8:testing")
+include(":instrumentation:rocketmq:rocketmq-clients:rocketmq-client-4.8:javaagent")
+include(":instrumentation:rocketmq:rocketmq-clients:rocketmq-client-4.8:library")
+include(":instrumentation:rocketmq:rocketmq-clients:rocketmq-client-4.8:testing")
 include(":instrumentation:runtime-metrics:javaagent")
 include(":instrumentation:runtime-metrics:library")
 include(":instrumentation:rxjava:rxjava-1.0:library")
